backend/app/requirements.txt
requests==2.31.0
pyotp==2.3.0
python-dotenv==0.15.0
cryptography==41.0.3
robin_stocks==2.1.6